custom_auth/requirements.txt
asgiref==3.9.2
bcrypt==5.0.0
Django==4.2.25
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
